+++ jvm-sandbox-inspector-agent/jvm-sandbox-inspector-agent-core/src/main/java/me/deipss/jvm/sandbox/inspector/agent/core/util/ConfigUtil.java
@@ -32,7 +32,15 @@ public class ConfigUtil {
         return config.getDataSendUrl();
     }
 
+    public static String getPluginsUrl(){
+        return config.getPluginsUrl();
+    }
+
     public static List<String> getEnablePlugins() {
         return config.getEnablePlugins();
+    }
+
+    public static List<String> getBizLoadClassRegexes() {
+        return config.getBizLoadClassRegexes();
     }
 }
