--- jvm-sandbox-inspector-agent/jvm-sandbox-inspector-agent-plugin/src/main/java/me/deipss/jvm/sandbox/inspector/agent/core/plugin/rocket/RocketMqConsumerPlugin.java
@@ -1,39 +1,0 @@
-package me.deipss.jvm.sandbox.inspector.agent.core.plugin.rocket;
-
-import com.alibaba.jvm.sandbox.api.event.Event;
-import com.alibaba.jvm.sandbox.api.listener.ext.EventWatchBuilder;
-import com.alibaba.jvm.sandbox.api.resource.ModuleEventWatcher;
-import me.deipss.jvm.sandbox.inspector.agent.api.Constant;
-import me.deipss.jvm.sandbox.inspector.agent.api.service.InvocationSendService;
-import me.deipss.jvm.sandbox.inspector.agent.api.spi.InspectorPlugin;
-import me.deipss.jvm.sandbox.inspector.agent.core.plugin.BasePlugin;
-import org.kohsuke.MetaInfServices;
-
-@MetaInfServices(InspectorPlugin.class)
-public class RocketMqConsumerPlugin extends BasePlugin {
-
-
-    @Override
-    public String identify() {
-        return Constant.ROCKET_MQ_CONSUMER;
-    }
-
-    @Override
-    public boolean entrance() {
-        return true;
-    }
-
-    @Override
-    public int watch(ModuleEventWatcher watcher, InvocationSendService invocationSendService) {
-        new EventWatchBuilder(watcher)
-                .onClass("org.apache.rocketmq.client.consumer.listener.MessageListenerConcurrently").includeBootstrap()
-                .includeSubClasses()
-                .onBehavior("consumeMessage")
-                .onClass("org.apache.rocketmq.client.consumer.listener.MessageListenerOrderly").includeBootstrap()
-                .includeSubClasses()
-                .onBehavior("consumeMessage")
-                .onWatch(new RocketMqConsumerEventListener(entrance(), identify(),invocationSendService), Event.Type.BEFORE, Event.Type.RETURN, Event.Type.THROWS);
-
-        return 0;
-    }
-}
