+++ jvm-sandbox-inspector-agent/jvm-sandbox-inspector-agent-plugins/pom.xml
@@ -9,47 +9,44 @@
         <version>${revision}</version>
     </parent>
 
-    <artifactId>jvm-sandbox-inspector-agent-core</artifactId>
-    <version>${revision}</version>
-
+    <artifactId>jvm-sandbox-inspector-agent-plugins</artifactId>
 
     <dependencies>
         <dependency>
             <groupId>com.alibaba.jvm.sandbox</groupId>
             <artifactId>sandbox-api</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>me.deipss</groupId>
             <artifactId>jvm-sandbox-inspector-agent-api</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>transmittable-thread-local</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>com.alibaba</groupId>
-            <artifactId>fastjson</artifactId>
+            <groupId>me.deipss</groupId>
+            <artifactId>jvm-sandbox-inspector-agent-core</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
             <artifactId>slf4j-api</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
             <groupId>ch.qos.logback</groupId>
             <artifactId>logback-classic</artifactId>
+            <scope>provided</scope>
         </dependency>
         <dependency>
-            <groupId>org.apache.commons</groupId>
-            <artifactId>commons-lang3</artifactId>
-        </dependency>
-        <dependency>
-            <groupId>org.apache.httpcomponents</groupId>
-            <artifactId>httpclient</artifactId>
+            <groupId>org.apache.dubbo</groupId>
+            <artifactId>dubbo</artifactId>
+            <version>2.7.4.1</version>
+            <scope>provided</scope>
         </dependency>
     </dependencies>
 
     <build>
-        <finalName>inspector-module</finalName>
+        <finalName>inspector-plugins</finalName>
         <plugins>
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
@@ -64,7 +61,7 @@
                             <descriptorRefs>
                                 <descriptorRef>jar-with-dependencies</descriptorRef>
                             </descriptorRefs>
-                            <finalName>inspector-module</finalName>
+                            <finalName>inspector-plugins</finalName>
                             <encoding>UTF-8</encoding>
                         </configuration>
                     </execution>
@@ -72,4 +69,5 @@
             </plugin>
         </plugins>
     </build>
+
 </project>
