+++ jvm-sandbox-inspector-agent/jvm-sandbox-inspector-agent-plugins/src/main/java/me/deipss/jvm/sandbox/inspector/agent/core/plugins/http/HttpInvocation.java
@@ -1,0 +1,17 @@
+package me.deipss.jvm.sandbox.inspector.agent.core.plugins.http;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.EqualsAndHashCode;
+import lombok.NoArgsConstructor;
+import me.deipss.jvm.sandbox.inspector.agent.api.domain.Invocation;
+import me.deipss.jvm.sandbox.inspector.agent.core.plugins.http.copier.HttpServletResponseCopier;
+
+@EqualsAndHashCode(callSuper = true)
+@AllArgsConstructor
+@NoArgsConstructor
+@Data
+public class HttpInvocation extends Invocation {
+
+    private transient HttpServletResponseCopier httpServletResponseCopier;
+}
