+++ jvm-sandbox-inspector-debug-consumer/pom.xml
@@ -86,10 +86,6 @@
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-spring-boot-starter</artifactId>
-        </dependency>
         <!--mq-->
         <dependency>
             <groupId>org.apache.rocketmq</groupId>
