+++ jvm-sandbox-inspector-debug-provider/pom.xml
@@ -87,10 +87,6 @@
             <groupId>org.apache.dubbo</groupId>
             <artifactId>dubbo</artifactId>
         </dependency>
-        <dependency>
-            <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-spring-boot-starter</artifactId>
-        </dependency>
         <!--mq-->
         <dependency>
             <groupId>org.apache.rocketmq</groupId>
